+++ src/lib/converter/index.ts
@@ -1,6 +1,6 @@
 export { Context } from "./context";
 export { Converter } from "./converter";
-
+export type { CommentParserConfig } from "./comments/index";
 export { convertDefaultValue, convertExpression } from "./convert-expression";
 
 import "./plugins/index";