+++ src/lib/models/reflections/abstract.ts
@@ -215,8 +215,8 @@ export interface TraverseCallback {
  * by the {@link DeclarationReflection} class.
  *
  * This base class exposes the basic properties one may use to traverse the reflection tree.
- * You can use the {@link ContainerReflection.children} and {@link parent} properties to walk the tree. The {@link groups} property
- * contains a list of all children grouped and sorted for being rendered.
+ * You can use the {@link ContainerReflection.children} and {@link parent} properties to walk the tree. The {@link ContainerReflection.groups} property
+ * contains a list of all children grouped and sorted for rendering.
  */
 export abstract class Reflection {
     /**